separate common function from run_load.go
