+++ load/run_load.go
@@ -1,22 +1,13 @@
 package load
 
 import (
-	"bytes"
-	"crypto/tls"
 	"fmt"
-	"hash"
 	"hash/fnv"
-	"io"
-	"io/ioutil"
 	"log"
 	"math"
 	"math/rand"
-	"net/http"
-	"net/http/httptrace"
-	"net/url"
 	"os"
 	"os/signal"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -29,214 +20,6 @@ import (
 	"github.com/buoyantio/slow_cooker/window"
 	"github.com/codahale/hdrhistogram"
 )
-
-// DayInMs 1 day in milliseconds
-const (
-	DayInMs                 int64  = 24 * 60 * 60 * 1000000
-	ServerBackendPrometheus string = "prometheus"
-	ServerBackendInfluxDB   string = "influxdb"
-)
-
-// MeasuredResponse holds metadata about the response
-// we receive from the server under test.
-type MeasuredResponse struct {
-	sz              uint64
-	code            int
-	latency         int64
-	timeout         bool
-	failedHashCheck bool
-	err             error
-}
-
-func newClient(
-	compress bool,
-	https bool,
-	noreuse bool,
-	maxConn int,
-) *http.Client {
-	tr := http.Transport{
-		DisableCompression:  !compress,
-		DisableKeepAlives:   noreuse,
-		MaxIdleConnsPerHost: maxConn,
-		Proxy:               http.ProxyFromEnvironment,
-	}
-	if https {
-		tr.TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
-	}
-	return &http.Client{Transport: &tr}
-}
-
-func sendRequest(
-	client *http.Client,
-	method string,
-	url *url.URL,
-	host string,
-	headers HeaderSet,
-	requestData []byte,
-	reqID uint64,
-	hashValue uint64,
-	checkHash bool,
-	hasher hash.Hash64,
-	received chan *MeasuredResponse,
-	bodyBuffer []byte,
-) {
-	req, err := http.NewRequest(method, url.String(), bytes.NewBuffer(requestData))
-	if err != nil {
-		fmt.Fprintln(os.Stderr, err.Error())
-		fmt.Fprintf(os.Stderr, "\n")
-	}
-	if host != "" {
-		req.Host = host
-	}
-	req.Header.Add("Sc-Req-Id", strconv.FormatUint(reqID, 10))
-	for k, v := range headers {
-		req.Header.Add(k, v)
-	}
-
-	var elapsed time.Duration
-	start := time.Now()
-
-	trace := &httptrace.ClientTrace{
-		GotFirstResponseByte: func() {
-			elapsed = time.Since(start)
-		},
-	}
-
-	req = req.WithContext(httptrace.WithClientTrace(req.Context(), trace))
-	response, err := client.Do(req)
-
-	if err != nil {
-		received <- &MeasuredResponse{err: err}
-	} else {
-		defer response.Body.Close()
-		if !checkHash {
-			if sz, err := io.CopyBuffer(ioutil.Discard, response.Body, bodyBuffer); err == nil {
-
-				received <- &MeasuredResponse{
-					sz:      uint64(sz),
-					code:    response.StatusCode,
-					latency: elapsed.Nanoseconds() / 1000000}
-			} else {
-				received <- &MeasuredResponse{err: err}
-			}
-		} else {
-			if bytes, err := ioutil.ReadAll(response.Body); err != nil {
-				received <- &MeasuredResponse{err: err}
-			} else {
-				hasher.Write(bytes)
-				sum := hasher.Sum64()
-				failedHashCheck := false
-				if hashValue != sum {
-					failedHashCheck = true
-				}
-				received <- &MeasuredResponse{
-					sz:              uint64(len(bytes)),
-					code:            response.StatusCode,
-					latency:         elapsed.Nanoseconds() / 1000000,
-					failedHashCheck: failedHashCheck}
-			}
-		}
-	}
-}
-
-func ExUsage(msg string, args ...interface{}) {
-	fmt.Fprintln(os.Stderr, fmt.Sprintf(msg, args...))
-	fmt.Fprintln(os.Stderr, "Try --help for help.")
-	os.Exit(64)
-}
-
-// CalcTimeToWait calculates how many Nanoseconds to wait between actions.
-func CalcTimeToWait(qps *int) time.Duration {
-	return time.Duration(int(time.Second) / *qps)
-}
-
-var reqID = uint64(0)
-
-var shouldFinish = false
-var shouldFinishLock sync.RWMutex
-
-// finishSendingTraffic signals the system to stop sending traffic and clean up after itself.
-func finishSendingTraffic() {
-	shouldFinishLock.Lock()
-	shouldFinish = true
-	shouldFinishLock.Unlock()
-}
-
-type HeaderSet map[string]string
-
-func (h *HeaderSet) String() string {
-	return ""
-}
-
-func (h *HeaderSet) Set(s string) error {
-	parts := strings.SplitN(s, ":", 2)
-	if len(parts) < 2 || len(parts[0]) == 0 {
-		return fmt.Errorf("Header invalid")
-	}
-	name := strings.TrimSpace(parts[0])
-	value := strings.TrimSpace(parts[1])
-	(*h)[name] = value
-	return nil
-}
-
-func LoadData(data string) []byte {
-	var file *os.File
-	var requestData []byte
-	var err error
-	if strings.HasPrefix(data, "@") {
-		path := data[1:]
-		if path == "-" {
-			file = os.Stdin
-		} else {
-			file, err = os.Open(path)
-			if err != nil {
-				fmt.Fprintf(os.Stderr, err.Error())
-				os.Exit(1)
-			}
-			defer file.Close()
-		}
-
-		requestData, err = ioutil.ReadAll(file)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, err.Error())
-			os.Exit(1)
-		}
-	} else {
-		requestData = []byte(data)
-	}
-
-	return requestData
-}
-
-// Sample Rate is between [0.0, 1.0] and determines what percentage of request bodies
-// should be checked that their hash matches a known hash.
-func ShouldCheckHash(sampleRate float64) bool {
-	return rand.Float64() < sampleRate
-}
-
-type RunLoadParams struct {
-	Qps                  int
-	Concurrency          int
-	Method               string
-	Interval             time.Duration
-	Noreuse              bool
-	Compress             bool
-	NoLatencySummary     bool
-	ReportLatenciesCSV   string
-	TotalRequests        uint64
-	Headers              HeaderSet
-	MetricAddr           string
-	HashValue            uint64
-	HashSampleRate       float64
-	DstURL               url.URL
-	Hosts                []string
-	RequestData          []byte
-	MetricsServerBackend string
-	InfluxUsername       string
-	InfluxPassword       string
-	InfluxDatabase       string
-	HistogramWindowSize  time.Duration
-}
 
 func RunLoad(params RunLoadParams) {
 	// Repsonse tracking metadata.
